# api-contracts/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/tenants:
    get:
      operationId: TenantsController_getTenants
      parameters: []
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantsResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      summary: Get all tenants
      tags:
        - Admin Tenants
    post:
      operationId: TenantsController_createTenant
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequestDto'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantsResponseDto'
        '400':
          description: Bad Request
        '409':
          description: Conflict - Tenant with the same name already exists
      summary: Create a new tenant
      tags:
        - Admin Tenants
  /admin/tenants/{id}/users:
    get:
      operationId: TenantsController_getTenantUsers
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    user_id:
                      type: string
                    role:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      summary: Get all users of a tenant
      tags:
        - Admin Tenants
  /admin/tenants/{id}:
    patch:
      operationId: TenantsController_updateTenant
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequestDto'
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantsResponseDto'
        '400':
          description: Bad Request
        '404':
          description: Tenant not found
      summary: Update tenant details
      tags:
        - Admin Tenants
    delete:
      operationId: TenantsController_deleteTenant
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Tenant deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTenantResponseDto'
        '404':
          description: Tenant not found
      summary: Delete a tenant
      tags:
        - Admin Tenants
  /alerts:
    get:
      operationId: AlertsController_getAlerts
      parameters:
        - name: from
          required: false
          in: query
          schema:
            type: string
        - name: to
          required: false
          in: query
          schema:
            type: string
        - name: gateway_id
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get alerts for the tenant in a time range
      tags:
        - Alerts
...
  /alerts/config/default:
    put:
      operationId: AlertsController_setDefaultAlertsConfig
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAlertsConfigDefaultRequestDto'
      responses:
        '200':
          description: ''
      summary: Set default alert configuration for the tenant
      tags:
        - Alerts
  /alerts/config/gateway/{gatewayId}:
    put:
      operationId: AlertsController_setGatewayAlertsConfig
      parameters:
        - name: gatewayId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGatewayAlertsConfigRequestDto'
      responses:
        '200':
          description: Alert configuration updated successfully
        '400':
          description: Bad Request
        '403':
          description: Gateway not associated with tenant of JWT
        '404':
          description: Gateway not found
      summary: Set alert configuration for a specific gateway
      tags:
        - Alerts
    delete:
      operationId: AlertsController_deleteGatewayAlertsConfig
      parameters:
        - name: gatewayId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Alert configuration deleted successfully
        '403':
          description: Gateway not associated with tenant of JWT
        '404':
          description: Gateway not found
      summary: Delete alert configuration for a specific gateway
      tags:
        - Alerts
  /api-clients:
    post:
      operationId: ApiClientController_createApiClient
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiClientRequestDto'
      responses:
        '400':
          description: Bad Request
        '409':
          description: API client with the same name already exists
      summary: Create a new API client for the tenant
      tags:
        - ApiClient
    get:
      operationId: ApiClientController_getApiClients
      parameters: []
      responses:
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      summary: Get all API clients for the tenant
      tags:
        - ApiClient
  /api-clients/{id}:
    delete:
      operationId: ApiClientController_deleteApiClient
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: API client deleted
        '404':
          description: API client not found
      summary: Delete an API client for the tenant
      tags:
        - ApiClient
...
  /gateways/{id}:
    get:
      operationId: GatewaysController_getGatewayById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayResponseDto'
      summary: Get gateway by id
      tags:
        - Gateways
    patch:
      operationId: GatewaysController_updateGateway
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGatewayRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateGatewayResponseDto'
      summary: Update gateway details
      tags:
        - Gateways
    delete:
      operationId: GatewaysController_deleteGateway
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Gateway deleted
      summary: Delete a gateway
      tags:
        - Gateways
...
info:
  title: NoTIP Management API
  description: NoTIP Management API OpenAPI specification
  version: 1.2.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    TenantsResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - active
            - suspended
        created_at:
          format: date-time
          type: string
      required:
        - id
        - name
        - status
        - created_at
    CreateTenantRequestDto:
      type: object
      properties:
        name:
          type: string
        admin_email:
          type: string
        admin_name:
          type: string
        admin_password:
          type: string
          example: Temp_1234_A!
      required:
        - name
        - admin_email
        - admin_name
        - admin_password
    UpdateTenantRequestDto:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum:
            - active
            - suspended
        suspension_interval_days:
          type: number
      required:
        - name
        - status
        - suspension_interval_days
    DeleteTenantResponseDto:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    SetAlertsConfigDefaultRequestDto:
      type: object
      properties:
        tenant_unreachable_timeout_ms:
          type: number
      required:
        - tenant_unreachable_timeout_ms
    SetGatewayAlertsConfigRequestDto:
      type: object
      properties:
        gateway_unreachable_timeout_ms:
          type: number
      required:
        - gateway_unreachable_timeout_ms
    CreateApiClientRequestDto:
      type: object
      properties: {}
...
    GatewayResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the gateway
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: User-defined name of the gateway
          example: Main Entrance Gateway
        status:
          type: string
          description: Current connectivity status
          enum:
            - gateway_online
            - gateway_offline
            - gateway_suspended
            - gateway_provisioning
          example: gateway_online
        last_seen_at:
          type: string
          description: Timestamp of the last heart-beat received from the gateway
          example: '2024-03-24T10:00:00Z'
          format: date-time
        provisioned:
          type: boolean
          description: Whether the gateway has been provisioned/activated
          example: true
        firmware_version:
          type: string
          description: Current firmware version installed on the device
          example: 1.2.3
        send_frequency_ms:
          type: number
          description: Configured data sending frequency in milliseconds
          example: 30000
      required:
        - id
        - name
        - status
        - last_seen_at
        - provisioned
        - firmware_version
        - send_frequency_ms
    UpdateGatewayRequestDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateGatewayResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - gateway_online
            - gateway_offline
            - gateway_suspended
            - gateway_provisioning
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - updated_at
...